Moved files to the right place
