+++ src/java/Controller/PDF.java
@@ -1,43 +1,42 @@
 package Controller;
 //You will need apache pdfbox, apache fontbox, apache commonsloggings, jdbc JARs
 
-import java.io.FileOutputStream;
-import java.io.OutputStream;
 import java.util.ArrayList;
+import java.util.List;
 import javax.sql.DataSource;
+import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDPage;
+import org.apache.pdfbox.pdmodel.PDPageContentStream;
+import org.apache.pdfbox.pdmodel.font.PDFont;
+import org.apache.pdfbox.pdmodel.font.PDType1Font;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-
-//pdf-box layout used under MIT license. Found at https://github.com/ralfstuckert/pdfbox-layout
-import rst.pdfbox.layout.elements.Document;
-import rst.pdfbox.layout.elements.Paragraph;
-import rst.pdfbox.layout.elements.VerticalSpacer;
-import rst.pdfbox.layout.elements.render.VerticalLayoutHint;
-import rst.pdfbox.layout.text.Alignment;
-import rst.pdfbox.layout.text.BaseFont;
 
 public class PDF {
     
     private static DataSource dataSource;
-    private NamedParameterJdbcTemplate njdbc;
 
-	private BaseFont font;
+    private PDDocument document = new PDDocument();
+	private PDFont font = PDType1Font.TIMES_ROMAN;
 	private PDPage page = new PDPage();
     private PDFinfo pdfinfo;
 
 	private String filePath = "C:/Users/syntel/Music/";//Path to save the pdf
+
+    private NamedParameterJdbcTemplate njdbc;
     
-    public PDF(){this.font = BaseFont.Times;
-}
+    public PDF(){}
     
 	public PDF(DataSource dataSource) {
-        this.font = BaseFont.Times;
         //super();
         System.out.println("Constructor Called");
         PDF.dataSource=dataSource;
-        njdbc = new NamedParameterJdbcTemplate(dataSource);
 	}
-   
+
+    @Override
+    public String toString() {
+        return "PDF{" + "document=" + document + ", font=" + font + ", page=" + page + ", pdfinfo=" + pdfinfo + ", filePath=" + filePath + ", njdbc=" + njdbc + '}';
+    }
+    
     public void setPdfinfo(PDFinfo pdfinfo) {
         this.pdfinfo = pdfinfo;
     }
@@ -61,6 +60,22 @@ public class PDF {
     public void setNjdbc(NamedParameterJdbcTemplate njdbc) {
         this.njdbc = njdbc;
     }
+
+    public PDDocument getDocument() {
+		return document;
+	}
+    
+	public void setDocument(PDDocument document) {
+		this.document = document;
+	}
+    
+	public PDFont getFont() {
+		return font;
+	}
+    
+	public void setFont(PDFont font) {
+		this.font = font;
+	}
     
 	public PDPage getPage() {
 		return page;
@@ -79,25 +94,35 @@ public class PDF {
 	}
     
 	public void generate(String empid) throws Exception{ //empid not implemented yet, should be able to get everything you will need from it
-        //define output vars
-        ArrayList<String> stream = new ArrayList();
-        ArrayList<String> foundations = new ArrayList();
-        ArrayList<String> specializations = new ArrayList();
-        ArrayList<String> domains = new ArrayList();         
-        String name, avgGrade, fGrade, sGrade, dGrade;
-        
-        //init config
+		System.out.println("Generating some stuff yo---------------------------------------------------------------------------------------------------");
+        List<String> stream = new ArrayList();
+            
+        String name;
+        String id;
+        String foundation;
+        String specialization;
+        String domain;
+        String avgGrade;
+        String fGrade;
+        String sGrade;
+        String dGrade;
+        String induction;
+        System.out.println("Connection Established");
         njdbc = new NamedParameterJdbcTemplate(dataSource);
         PDFinfo info = new PDFinfo(dataSource.getConnection());
         info.setDataSource(dataSource);
         info.setNjdbc(njdbc);
-        
-        if(!info.getAllIds().contains(empid)){ //this is O(n^2) methinks. make the query return one think and do bool return.
-            System.out.println("Employee ID not found. PDF has not been generated.");
-            return;
-        }
-
-        //populate output variables
+        if(info.getAllIds().contains(empid)){
+        System.out.println("Connection Established");
+		document = new PDDocument();
+		font = PDType1Font.TIMES_ROMAN;
+		page = new PDPage();
+		document.addPage( page );
+		ArrayList <String>foundations = new ArrayList();
+        ArrayList <String>specializations = new ArrayList();
+        ArrayList <String>domains = new ArrayList();
+		PDPageContentStream contentStream = new PDPageContentStream(document, page);
+                
         name=info.getEmployeeName(empid);
         stream=info.getStreamIDName(empid);
         avgGrade=info.getAverageScoresByEmployeeID(empid);
@@ -107,83 +132,57 @@ public class PDF {
         fGrade=info.getAverageScoresByFoundationEmployeeID(empid);
         sGrade=info.getAverageScoresBySpecializationEmployeeID(empid);
         dGrade=info.getAverageScoresByProcessDomainEmployeeID(empid);
-		
-        Document document = new Document(40, 50, 40, 60);
-        float linspace = 4;
-        float sectionBreak = 6.5f;
-        
-        Paragraph title = new Paragraph();
-        title.addMarkup("*PERFORMICA REPORT*", 20, font);
-       	document.add(title, VerticalLayoutHint.CENTER);
-        document.add(new VerticalSpacer(5));
-   
-        Paragraph emp = new Paragraph();
-        emp.addMarkup( "*NAME*: " + name + "            " +
-                "*EMPLOYEE ID*: " + empid, 14, font);
-        emp.setAlignment(Alignment.Center);
-        document.add(emp);
-        
-        document.add(new VerticalSpacer(6.5f));
-        
-        Paragraph strm = new Paragraph();
-        strm.addMarkup("*INDUCTION*: " + stream.get(0) + " - " + stream.get(1), 14, font);
-        strm.setAlignment(Alignment.Center);
-        document.add(strm);
-		
-        document.add(new VerticalSpacer(15.5f));
-        
-        Paragraph p1 = new Paragraph();
-        p1.addMarkup("__*Training Modules Completed*__", 12, font);
-        document.add(p1);
-        document.add(new VerticalSpacer(linspace));
-        
-        Paragraph found = new Paragraph();
-        found.addMarkup("*Foundations*:  ", 12, font);
-        for(int i = 0; i < foundations.size() - 2; i+=2){
-            found.addMarkup(foundations.get(i) + ",  ", 12, font);
+		contentStream.setLeading(14.5f);
+		contentStream.beginText();
+		contentStream.setFont( font, 76 );
+		contentStream.newLineAtOffset(100, 700);
+        contentStream.showText("LOGO HERE");
+        contentStream.newLine();
+        contentStream.setFont( font, 12);
+		contentStream.showText( "Training Report: "+name+"            ");//NAME select name from employees where employee_id=empid?
+        contentStream.showText("Emp ID: "+empid);//Just use empid param
+		contentStream.newLine();
+		contentStream.showText("Induction: "+stream.toString());//Stream Code + Stream Name select s.stream_id, s.stream_name from Stream s, Class c, Employees e where s.stream_id=c.stream_id and c.class_id=e.class_id and e.employee_id='ab'
+		contentStream.newLine();
+        contentStream.newLine();
+        contentStream.showText("Foundation: ");//select distinct c.course_name from 
+        contentStream.newLine();
+        for(int i = 0; i < foundations.size(); i+=2){
+            contentStream.showText(foundations.get(i)+" "+foundations.get(i+1));
+            contentStream.newLine();
         }
-        found.addMarkup(foundations.get(foundations.size() - 2), 12, font);
-        document.add(found);
-        document.add(new VerticalSpacer(linspace));
-
-        Paragraph spec = new Paragraph();
-        spec.addMarkup("*Specializations*: ", 12, font);
-        for(int i = 0; i < specializations.size() - 2; i+=2){
-            spec.addMarkup(specializations.get(i) + ",  ", 12, font);
+        contentStream.newLine();
+        contentStream.showText("Specialization: ");
+        contentStream.newLine();
+          for(int i = 0; i < specializations.size(); i+=2){
+            contentStream.showText(specializations.get(i)+" "+specializations.get(i+1));
+            contentStream.newLine();
         }
-        spec.addMarkup(specializations.get(specializations.size() - 2), 12, font);  
-        document.add(spec);
-        document.add(new VerticalSpacer(linspace));
-        
-        Paragraph pd = new Paragraph();
-        pd.addMarkup("*Process / Domain*: ", 12, font);
-        for(int i = 0; i < domains.size() - 2; i+=2){
-            pd.addMarkup(domains.get(i) + ",  ", 12, font);
+        contentStream.newLine();
+        contentStream.showText("Process/Domain: ");
+        contentStream.newLine();
+          for(int i = 0; i < domains.size(); i+=2){
+            contentStream.showText(domains.get(i)+" "+domains.get(i+1));
+            contentStream.newLine();
         }
-        pd.addMarkup(domains.get(domains.size() - 2), 12, font);  
-        document.add(pd);
-        
-        document.add(new VerticalSpacer(sectionBreak));
-        
-        Paragraph p2 = new Paragraph();
-        p2.addMarkup("__*Performence Report*__\nThis is where we would put the bargraph if we had it", 12, font);
-        document.add(new VerticalSpacer(linspace));
-        document.add(p2);
-        document.add(new VerticalSpacer(350));
-        
-        String gradeNumbers = "Overall Grade: "+avgGrade+"  |  "
-                + "Foundation Grade: "+fGrade+"  |  "
-                + "Specialization Grade: "+sGrade+"  |  "
-                + "Process/Domain Grade: "+dGrade+"   ";
-        Paragraph grades = new Paragraph();
-        grades.addMarkup(gradeNumbers, 12, font);
-        document.add(grades);
-        document.add(new VerticalSpacer(linspace));
-        
-        final OutputStream outputStream = new FileOutputStream(
-		filePath + "/PDF.pdf");
-        document.save(outputStream);
-    
+        contentStream.newLine();
+        contentStream.showText("Overall Grade:  "+avgGrade+"   ");//select avg(e.scores) from Employees_take_Modules e, Modules m where m.module_id=e.module_id and e.employee_id=?
+        contentStream.showText("Foundation Grade: "+fGrade+"   ");//select avg(e.scores) from Employees_take_Modules e, Modules m where m.category='Foundation' and m.module_id=e.module_id and e.employee_id=?
+        contentStream.showText("Specialization Grade: "+sGrade+"   ");//select avg(e.scores) from Employees_take_Modules e, Modules m where m.category='Specialization' and m.module_id=e.module_id and e.employee_id=?
+        contentStream.showText("Process/Domain Grade: "+dGrade+"   ");//select avg(e.scores) from Employees_take_Modules e, Modules m where m.category='ProcessDomain' and m.module_id=e.module_id and e.employee_id=?
+		contentStream.endText();
+		contentStream.close();
+        System.out.println("PDF Finished Writing");
+		document.save(filePath + empid+".pdf"); //There is an issue here to work out
+        System.out.println("PDF Saved");
+		document.close();
+		System.out.println("PDF Closed");
+                
+        }
+        else{
+            System.out.println("Employee ID not found. PDF has not been generated.");
+        }
         //BarChartEx bc = new BarChartEx();
-	}	
+	}
+		
 }
