+++ src/java/Controller/PDFinfo.java
@@ -5,6 +5,7 @@ import java.sql.ResultSet;
 import java.sql.Statement;
 import java.util.ArrayList;
 import javax.sql.DataSource;
+//import java.lang.String;
 import java.util.List;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
@@ -24,6 +25,21 @@ public class PDFinfo {
     }
 	
     public PDFinfo() {
+    }
+
+    /**
+    * Get employee ID of all employees
+    * @return ArrayList of all employee IDs
+    * @throws java.lang.Exception
+    */
+    public ArrayList<String> getAllIds() throws Exception{
+        ArrayList<String> list = new ArrayList();
+        Statement s1= c1.createStatement();
+        ResultSet r1=s1.executeQuery("select employee_id from employees");
+        while(r1.next()){
+            list.add(r1.getString(1));
+        }
+        return list;
     }
     
     public void setC1(Connection c1) {
@@ -50,28 +66,14 @@ public class PDFinfo {
         return dataSource;
     }
 	 
-    /**
-    * Get employee ID of all employees
-    * @return ArrayList of all employee IDs
-    * @throws java.lang.Exception
-    */
-    public ArrayList<String> getAllIds() throws Exception{
-        ArrayList<String> list = new ArrayList();
-        Statement s1= c1.createStatement();
-        ResultSet r1=s1.executeQuery("select employee_id from employees");
-        while(r1.next()){
-            list.add(r1.getString(1));
-        }
-        return list;
-    } 
-    
 	/**
     * Get employee name for a given Employee ID
-	* @param empID - the employee id of the employee
+	* @param empID - the employeee id of the employee
     * @return employee name
     * @throws java.lang.Exception
     */
     public String getEmployeeName(String empID) throws Exception{
+        System.out.println("Getting Employee Name");
         String SQL = "SELECT name FROM Employees WHERE employee_id = :id";
 
         SqlParameterSource namedParams= new MapSqlParameterSource("id",empID);
@@ -81,11 +83,11 @@ public class PDFinfo {
 	
 	/**
     * Get stream ID and name of streams employee has undertaken
-    * @param empID - the employee id of the employee
+    * @param empID - the employeee id of the employee
     * @return a list of strings, formatted as [id1, name1, id2, name2, ...]
     * @throws java.lang.Exception
     */
-    public ArrayList<String> getStreamIDName(String empID) throws Exception{
+    public List<String> getStreamIDName(String empID) throws Exception{
 		ArrayList<String> list = new ArrayList();
         Statement s1 = c1.createStatement();
 		ResultSet r1=s1.executeQuery("select s.stream_id, s.stream_name from Stream s, Class c, Employees e where s.stream_id=c.stream_id and c.class_id=e.class_id and e.employee_id='"+empID+"'");
@@ -102,8 +104,8 @@ public class PDFinfo {
 	
     /**
     * Get overall average score of an employee
-    * @param empID - the employee id of the employee
-    * @return a string representation of the average score
+    * @param empID - the employeee id of the employee
+    * @return a string representaiton of the average score
     * @throws java.lang.Exception
     */
 	public String getAverageScoresByEmployeeID(String empID) throws Exception{
@@ -117,7 +119,7 @@ public class PDFinfo {
     /**
     * Get average foundation score of an employee
     * @param empID - the employee id of the employee
-    * @return a string representation of the average score
+    * @return a string representaiton of the average score
     * @throws java.lang.Exception
     */
 	public String getAverageScoresByFoundationEmployeeID(String empID) throws Exception{
@@ -129,9 +131,9 @@ public class PDFinfo {
 	}
 	
     /**
-    * Get average spec score of an employee
-    * @param empID - the employee id of the employee
-    * @return a string representation of the average score
+    * Get average specilization score of an employee
+    * @param empID - the employeee id of the employee
+    * @return a string representaiton of the average score
     * @throws java.lang.Exception
     */
 	public String getAverageScoresBySpecializationEmployeeID(String empID) throws Exception{
@@ -144,8 +146,8 @@ public class PDFinfo {
 	
     /**
     * Get average process/domain score of an employee
-    * @param empID - the employee id of the employee
-    * @return a string representation of the average score
+    * @param empID - the employeee id of the employee
+    * @return a string representaiton of the average score
     * @throws java.lang.Exception
     */
 	public String getAverageScoresByProcessDomainEmployeeID(String empID) throws Exception{
@@ -158,8 +160,8 @@ public class PDFinfo {
 	
     /**
     * Get list of foundation scores of an employee
-    * @param empID - the employee id of the employee
-    * @return a string representation of the average score
+    * @param empID - the employeee id of the employee
+    * @return a string representaiton of the average score
     * @throws java.lang.Exception
     */
 	public ArrayList<String> getModScoreByFoundation(String empID) throws Exception{
@@ -179,8 +181,8 @@ public class PDFinfo {
 	
     /**
     * Get list of spec scores of an employee
-    * @param empID - the employee id of the employee
-    * @return a string representation of the average score
+    * @param empID - the employeee id of the employee
+    * @return a string representaiton of the average score
     * @throws java.lang.Exception
     */
 	public ArrayList<String> getModScoreBySpecialization(String empID) throws Exception{
@@ -200,8 +202,8 @@ public class PDFinfo {
 	
     /**
     * Get list of p&d scores of an employee
-    * @param empID - the employee id of the employee
-    * @return a string representation of the average score
+    * @param empID - the employeee id of the employee
+    * @return a string representaiton of the average score
     * @throws java.lang.Exception
     */
 	public ArrayList<String> getModScoreByProcessDomain(String empID) throws Exception{
